refactor: unify prefixed number readers
